backend/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.1
pytz==2023.3
Werkzeug==2.3.7
openpyxl==3.1.2
pip install flask-jwt-extended

pip install flask-jwt-extended passlib bcrypt
